+++ .github/workflows/docker-build-push.yml
@@ -32,7 +32,7 @@ jobs:
 
             - name: Extract metadata (tags, labels) for Docker
               id: meta
-              uses: docker/metadata-action@v4
+              uses: docker/metadata-action@v5
               with:
                 images: ${{ secrets.DOCKER_USERNAME }}/react-router-backend
                 tags: |
