+++ backend/requirements.txt
@@ -7,20 +7,20 @@ PyJWT==2.5.0
 black==24.4.2
 flake8==7.0.0
 isort==5.13.2
-mypy==1.5.1
+mypy==1.10.1
 pre-commit==3.7.1
-sqlalchemy==2.0.27
+sqlalchemy==2.0.31
 sqlmodel==0.0.19
 psycopg==3.1.18
 psycopg-binary==3.2.1
 psycopg-pool==3.2.2
 psycopg2-binary==2.9.9
 Mako==1.3.5
-alembic==1.13.1
+alembic==1.13.2
 tenacity==8.2.3
 python-multipart==0.0.9
-uvicorn==0.27.1
-email-validator==2.1.1
+uvicorn==0.30.1
+email-validator==2.2.0
 pydantic-settings==2.2.1
 pytest>=7.4.2,<9.0.0
 httpx>=0.25.0,<0.26
